+++ db/sqlc/querier.go
@@ -12,11 +12,11 @@ import (
 type Querier interface {
 	CreateOwner(ctx context.Context, arg CreateOwnerParams) (sql.Result, error)
 	CreateOwnerPet(ctx context.Context, arg CreateOwnerPetParams) (sql.Result, error)
-	DeleteOwner(ctx context.Context, id int64) error
+	DeleteOwner(ctx context.Context, id int32) error
 	DeleteOwnerPet(ctx context.Context, arg DeleteOwnerPetParams) error
-	GetOwner(ctx context.Context, id int64) (GetOwnerRow, error)
+	GetOwner(ctx context.Context, id int32) (GetOwnerRow, error)
 	GetOwnerPet(ctx context.Context, arg GetOwnerPetParams) (Pet, error)
-	ListOwnerPets(ctx context.Context, ownerID sql.NullInt64) ([]Pet, error)
+	ListOwnerPets(ctx context.Context, ownerID sql.NullInt32) ([]Pet, error)
 	ListOwners(ctx context.Context) ([]ListOwnersRow, error)
 }
